Made changes to the D4R algorithm as discussed with Greg on Feb 17, 2010. We modified the algorithm such that the field that maps to M&M's priority field is unique by using the key as a tie breaker.

+++ libraries/D4R/D4RNode.h
@@ -72,7 +72,7 @@ namespace D4R {
     class Node {
     public:
 
-        Node(unsigned long long key);
+        Node(uint64_t key);
 
         virtual ~Node();
